feat: implement endpoints for posts

+++ src/main/java/org/example/post/Post.java
@@ -1,5 +1,5 @@
 package org.example.post;
 
-public record Post(Integer id, String title, String body) {
+public record Post(Long id, String title, String body) {
 
 }
